+++ provider_counter/.metadata
@@ -4,7 +4,7 @@
 # This file should be version controlled and should not be manually edited.
 
 version:
-  revision: 5391447fae6209bb21a89e6a5a6583cac1af9b4b
-  channel: stable
+  revision: 7a4c33425ddd78c54aba07d86f3f9a4a0051769b
+  channel: beta
 
 project_type: app
