+++ provider_counter/android/app/src/main/res/values/styles.xml
