+++ provider_counter/android/build.gradle
