+++ provider_counter/android/gradle.properties
