+++ provider_counter/android/gradle/wrapper/gradle-wrapper.properties
