+++ provider_counter/android/settings.gradle
