+++ provider_counter/ios/Flutter/AppFrameworkInfo.plist
