+++ provider_counter/ios/Runner/Assets.xcassets/AppIcon.appiconset/Contents.json
