+++ provider_counter/ios/Runner/Assets.xcassets/LaunchImage.imageset/Contents.json
