+++ provider_counter/ios/Runner/Assets.xcassets/LaunchImage.imageset/LaunchImage@3x.png
