+++ provider_counter/pubspec.yaml
@@ -1,8 +1,8 @@
-name: scoped_model_counter
+name: provider_counter
 description: >
-  The starter Flutter application, but using ScopedModel to manage state.
+  The starter Flutter application, but using Provider to manage state.
 
-version: 1.0.0+1
+version: 1.0.0
 
 environment:
   sdk: ">=2.0.0-dev.68.0 <3.0.0"
